refactor: move reducer helpers
